Mapeamento com DTOs e uso de data anotations

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,10 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using ProEventos.Application.Contratos;
+using ProEventos.Application.Dtos;
 using ProEventos.Domain;
 namespace ProEventos.API.Controllers
 {
@@ -26,7 +25,10 @@ namespace ProEventos.API.Controllers
             try
             {
                  var eventos = await _eventoService.GetAllEventosAsync(true);
-                 if (eventos == null) return NotFound("Nenhum evento encontrado.");
+                 if (eventos == null) return NoContent();
+
+                
+
                  return Ok(eventos);
             }
             catch (Exception ex)
@@ -42,7 +44,7 @@ namespace ProEventos.API.Controllers
             try
             {
                  var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                 if (evento == null) return NotFound("Evento por Id não encontrado.");
+                 if (evento == null) return NoContent();
                  return Ok(evento);
             }
             catch (Exception ex)
@@ -58,7 +60,7 @@ namespace ProEventos.API.Controllers
             try
             {
                  var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                 if (evento == null) return NotFound("Eventos por tema não encontrado.");
+                 if (evento == null) return NoContent();
                  return Ok(evento);
             }
             catch (Exception ex)
@@ -69,12 +71,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                  var evento = await _eventoService.AddEvento(model);
-                 if (evento == null) return BadRequest("Erro ao tentar adicionar eventos.");
+                 if (evento == null) return NoContent();
                  return Ok(evento);
             }
             catch (Exception ex)
@@ -85,12 +87,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
                  var evento = await _eventoService.UpdateEvento(id, model);
-                 if (evento == null) return BadRequest("Erro ao tentar atualizar eventos.");
+                 if (evento == null) return NoContent();
                  return Ok(evento);
             }
             catch (Exception ex)
@@ -105,9 +107,11 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                 if (evento == null) return NoContent();
                  return await _eventoService.DeleteEvento(id) ?
                         Ok("Evento Deletado.") :
-                        BadRequest("Evento não deletado.");
+                        throw new Exception("Ocorreu um erro desconhecido ao tentar deletar evento");
                  
             }
             catch (Exception ex)
